requirements.txt
python-telegram-bot>=13.7
firebase-admin>=5.0.1
python-i18n>=0.3.9
